requirements.txt
python-dotenv
langchain_community 
langchain
langchain_openai
langchain-groq
langchain_chroma
pypdf